% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhatKindOfTriangle.R
\name{WhatKindOfTriangle}
\alias{WhatKindOfTriangle}
\title{Finding out the kind of triangle which you are having.}
\usage{
WhatKindOfTriangle(a, b, c)
}
\description{
This function will tell you the kind of triangle which you are having by using 'Pythagorean theorem'.
But I found out that the code 'sqrt(2)^2-2' return 4.440892e-16 instead of 0.
That is the reason I made the code 'xa <- round(a^2) ~ xc <- round(c^2)' in this function.
And for example, if lengths of triangles's each side are 1.6, 1.6, sqrt(5.12), This triangle must be right-angled triangle.
But this function will return "This is isosceles,acute-angled triangle."
Thus, I will limit the kind of values of triangles's each side.
Only when these values are integer or the form 'sqrt(n)'(n must be natural number.), This function will be executed correctly.
}
\examples{

WhatKindOfTriangle(2,2,sqrt(8))
}
